src/requirements.txt
Django==1.10.6
django-jet==1.0.5
django-allauth==0.31.0
django-pagedown==0.1.3
django-ckeditor==5.2.2
oauthlib==2.0.2
pep8==1.7.0
Pillow==4.0.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
